+++ src/components/auth/github-login.js
@@ -1,10 +1,6 @@
 /** @jsx jsx */
 import { jsx } from 'theme-ui';
 import React from 'react';
-import Twitter from '../svg/twitter';
-
-import jwt_decode from 'jwt-decode';
-import { UserContext } from '../../context/user-context';
 import LogoButton from '../common/logo-button';
 import GitHub from './../svg/github';
 import { DevCardAuthContext, DevCardDispatchContext, DevCardStateContext } from '../../context/devcard-context';
