+++ src/components/hub/dev-card/images/profile-dropzone.js
@@ -3,18 +3,21 @@ import { jsx } from 'theme-ui';
 import React from 'react';
 import { useDropzone } from 'react-dropzone';
 import { useUpload } from 'use-cloudinary';
-import Loading from './../svg/loading';
-import Error from './../svg/error';
-import Cloud from '../svg/cloud';
-import { DevCardStateContext, DevCardDispatchContext } from '../../context/devcard-context';
+import Loading from '../../../svg/loading';
+import { DevCardStateContext, DevCardDispatchContext } from '../../../../context/devcard-context';
+import Cloud from '../../../svg/cloud';
 
-const CoverDropzone = () => {
+const ProfileDropzone = () => {
   const { upload, data, isLoading, isError, error } = useUpload({ endpoint: '/.netlify/functions/upload' });
   const state = React.useContext(DevCardStateContext);
   const dispatch = React.useContext(DevCardDispatchContext);
 
   React.useEffect(() => {
-    if (data && data.url) dispatch({ type: 'coverPreviewUrl', payload: data.url });
+    console.log('username: ', state.name);
+  }, []);
+
+  React.useEffect(() => {
+    if (data && data.url) dispatch({ type: 'profilePreviewUrl', payload: data.url });
   }, [data]);
 
   const onDrop = (acceptedFiles) => {
@@ -32,19 +35,19 @@ const CoverDropzone = () => {
     blobToBase64(acceptedFiles[0]).then((res) => {
       const name = state.cloudinaryFolderName ? state.cloudinaryFolderName : state.name;
       console.log({ name });
-      dispatch({ type: 'coverBase64Image', payload: res });
-      dispatch({ type: 'acceptedCoverFiles', payload: acceptedFiles[0] });
+      dispatch({ type: 'profileBase64Image', payload: res });
+      dispatch({ type: 'acceptedProfileFiles', payload: acceptedFiles[0] });
       return upload({
         // We pass the whole base64 string including the data:image tag
         file: res,
         uploadOptions: {
           // Get rid of the spaces in the name and attach the file path, giving the user its own folder with their name and their image inside
-          public_id: `${name.replace(/\s/g, '')}-cover/${acceptedFiles[0].path}`,
+          public_id: `${name.replace(/\s/g, '')}-profile/${acceptedFiles[0].path}`,
           tags: [],
           eager: [
             {
-              width: 1500,
-              height: 500,
+              width: 400,
+              height: 400,
               crop: 'fill',
             },
           ],
@@ -62,7 +65,7 @@ const CoverDropzone = () => {
     return (
       <div
         sx={{
-          gridArea: 'coverDropzone',
+          gridArea: 'profileDropzone',
           display: 'flex',
           height: 'auto',
           minHeight: 240,
@@ -83,7 +86,7 @@ const CoverDropzone = () => {
   return (
     <div
       sx={{
-        gridArea: 'coverDropzone',
+        gridArea: 'profileDropzone',
         display: 'flex',
         height: 'auto',
         border: 'solid 2px',
@@ -129,7 +132,7 @@ const CoverDropzone = () => {
                 fontFamily: 'heading',
               }}
             >
-              Upload your <strong>COVER</strong> image
+              Upload your <strong>PROFILE</strong> image
             </p>
             <em
               sx={{
@@ -149,7 +152,7 @@ const CoverDropzone = () => {
                 fontWeight: 700,
               }}
             >
-              Recommended size is 1500 x 500
+              The GitHub API doesn't allow you to update your profile picture
             </em> */}
           </div>
         </div>
@@ -177,4 +180,4 @@ const CoverDropzone = () => {
   );
 };
 
-export default CoverDropzone;
+export default ProfileDropzone;
