+++ src/components/hub/dev-card/social-checkboxes/checkbox.js
@@ -9,7 +9,7 @@ const Checkbox = ({ type, checked, onCheckboxChange, disabled, comingSoon }) => 
         cursor: 'pointer',
         position: 'relative',
       }}
-      htmlFor=""
+      htmlFor={type}
     >
       <input
         disabled={disabled}
@@ -25,16 +25,6 @@ const Checkbox = ({ type, checked, onCheckboxChange, disabled, comingSoon }) => 
           outline: 'none',
           backgroundColor: disabled ? 'disabled' : 'background',
           cursor: disabled ? null : 'pointer',
-          // ':after': {
-          //   content: comingSoon ? "'coming soon'" : "''",
-          //   top: 30,
-          //   left: 0,
-          //   position: 'absolute',
-          //   color: 'accent',
-          //   fontFamily: 'heading',
-          //   fontWeight: 600,
-          //   textTransform: 'uppercase',
-          // },
           ':checked': {
             borderColor: 'background',
             backgroundColor: 'background',
